MH template and caisemh
resolves #2088
also changed Kuimh to use template
